Crate icu_data_provider

Modules

decimal

Structs

Request
A struct to request a certain hunk of data from a data provider.

Response
A response object containing a data hunk ("payload").

ResponseBuilder
Builder class used to construct a Response.

Enums

Category
A top-level collection of related data keys.

Key
A specific key within a category.

PayloadError

ResponseError

Traits

DataProvider
An abstract data providewr that takes a request object and returns a response with a payload. Lifetimes: