Module icu_data_provider::prelude

Re-exports

pub use crate::icu_data_key;

Structs

DataEntry
A variant and language identifier, used for requesting data from a DataProvider.

DataKey
A category, subcategory, and version, used for requesting data from a DataProvider.

DataRequest
A struct to request a certain hunk of data from a data provider.

DataResponse
A response object containing a data hunk ("payload").

DataResponseBuilder
Builder class used to construct a Response.

Enums

DataCategory
A top-level collection of related data keys.

DataError

Traits

DataProvider
An abstract data provider that takes a request object and returns a response with a payload. Lifetimes: